refactor: Layout 컴포넌트의 cartItems fetch 로직을 useCartHandler 훅으로 이동
- 데이터 fetch 역할을 커스텀 훅들이 담당하도록 하여 코드 구조의 일관성 개선
