+++ src/pages/Layout.tsx
@@ -2,34 +2,20 @@ import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
 import Header from '../components/Header';
 import { CartProvider } from '../context/CartContext';
-import { useEffect, useState } from 'react';
-import tryApiCall from '../util/tryApiCall';
-import { getCartItems } from '../services/cartItemServices';
-import { CartItemType } from '../types/data';
+import { useState } from 'react';
 import { ErrorMessageProvider } from '../context/ErrorMessageContext';
+import useCartHandler from '../hooks/useCartHandler';
 
 const Layout = () => {
-  const [cartItemsIds, setCartItemsIds] = useState<number[]>([]);
   const [errorMessage, setErrorMessage] = useState('');
-
-  const handleAddCartItemsIds = (id: number) => {
-    setCartItemsIds((prev: number[]) => [...prev, id]);
-  };
-
-  const handleRemoveCartItemsIds = (id: number) => {
-    setCartItemsIds((prev: number[]) => prev.filter((itemId: number) => itemId !== id));
-  };
 
   const handleErrorMessage = (errorMessage: string) => {
     setErrorMessage(errorMessage);
   };
 
-  useEffect(() => {
-    (async () => {
-      const items = await tryApiCall(getCartItems, handleErrorMessage);
-      setCartItemsIds(items.map((item: CartItemType) => item.product.id));
-    })();
-  }, []);
+  const { cartItemsIds, handleAddCartItemsIds, handleRemoveCartItemsIds } = useCartHandler({
+    handleErrorMessage,
+  });
 
   return (
     <ErrorMessageProvider errorMessage={errorMessage} handleErrorMessage={handleErrorMessage}>
